# docs/openapi-draft.yaml
openapi: 3.1.0
info:
  title: XivDaily API Draft
  version: 0.1.0
  description: >
    XivDaily 当前后端接口草案。路径与字段以仓库内真实 FastAPI 实现为准。
servers:
  - url: http://127.0.0.1:8000
    description: 本机开发环境
...
paths:
...
  /zotero/sync/{paper_id}:
    post:
      tags: [Zotero]
      summary: 同步单篇论文到 Zotero
      operationId: syncPaperToZotero
      parameters:
        - in: path
          name: paper_id
          required: true
          schema:
            type: string
          description: 论文唯一标识
      responses:
        "200":
          description: 同步状态已返回
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoteroSyncResult"
  /zotero/exports/bibtex:
    post:
      tags: [Zotero]
      summary: 批量导出 BibTeX
      operationId: exportBibtex
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BibtexExportRequest"
      responses:
        "200":
          description: 导出成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BibtexExportResponse"
components:
  schemas:
...
    ZoteroSyncResult:
      type: object
      required:
        [paper_id, status, synced_at, target_collection_name, target_collection_status, visibility_status]
      properties:
        paper_id:
          type: string
        status:
          type: string
          enum: [synced, failed]
        zotero_item_key:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
        synced_at:
          type: string
          format: date-time
        library_type:
          type: string
          nullable: true
        user_id:
          type: string
          nullable: true
        target_collection_name:
          type: string
        target_collection_key:
          type: string
          nullable: true
        target_collection_status:
          type: string
          enum: [not_configured, ready, created, error]
        visibility_status:
          type: string
          enum: [verified, missing_from_collection, unverified, not_checked]
        visibility_message:
          type: string
          nullable: true
    BibtexExportRequest:
      type: object
      required: [paper_ids]
      properties:
        paper_ids:
          type: array
          items:
            type: string
    BibtexExportResponse:
      type: object
      required: [content, exported_count]
      properties:
        content:
          type: string
        exported_count:
          type: integer